+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Decouple `Service` from controllers using an slice of controllers.
+- Retry failed ARM deployments regardless of the checksum check.
 
 ## [4.2.0] - 2020-07-28